refactor : 이전 커밋 pull 하면서 EventRepository 패키지 변경으로 인한 import 변경
#83
