--- backend/emm-sale/src/main/java/com/emmsale/event/domain/EventRepository.java
@@ -1,7 +1,0 @@
-package com.emmsale.event.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface EventRepository extends JpaRepository<Event, Long> {
-
-}
